[RoboCIn] Refactor: replace BaseReplayBuffer with ReplayBuffer and remove unused files
